refactored group delete

+++ apps/rahat-community/src/app/groups/group.service.ts
@@ -83,7 +83,7 @@ export class GroupService {
     });
   }
 
-  async remove(uuid: string) {
+  async remove(uuid: string, deleteBeneficiaryFlag: boolean) {
     // get relevant informationn from the group table
     const getInfo = await this.prisma.group.findUnique({
       where: {
@@ -104,18 +104,21 @@ export class GroupService {
       await this.prisma.$transaction(async (prisma) => {
         for (const item of getInfo.beneficiariesGroup) {
           // first delete from the combine table (tbl_beneficiary_groups)
+
           await prisma.beneficiaryGroup.deleteMany({
             where: {
               beneficiaryId: item.beneficiaryId,
             },
           });
 
-          // delete beneficiary from the beneficiary table (tbl_beneficiaries)
-          await prisma.beneficiary.delete({
-            where: {
-              id: item?.beneficiaryId,
-            },
-          });
+          if (deleteBeneficiaryFlag) {
+            // delete beneficiary from the beneficiary table (tbl_beneficiaries)
+            await prisma.beneficiary.delete({
+              where: {
+                id: item?.beneficiaryId,
+              },
+            });
+          }
         }
       });
     }
